Elastic tabs: changed memory management.
Before, each ElasticTabs kept the line numbers of its first and last line, and
was deleted when the last line was.  This can be made to work, but is too
clever and fragile, because it requires diligence about keeping those line
numbers updated (*and* the dirty line numbers it also used to keep) when lines
are deleted.  Already there were observed and suspected bugs.

Now it uses simple reference-counting.  Diligence is still required, but it's a
lot easier.

Also, it no longer tries to optimize recalculation of the column widths when
there were only character additions, not deletions.  This may have been a
premature optimization -- a full recalculation of a column widths for the
entire History (currently fixed at 10000 lines) took no more than 10ms, once
per draw.