els, end-tabs: Trying to deal with broken pipes better.
Broken pips occur when piping through "head", for instance.  "els" can't output
the end-of-elastic-tabs code, because the pipe is already gone by the time it
finds out.  "end-tabs" is a way (albeit clunky) to deal with that.  And "els"
no longer complains about the broken pipe on stderr.

"els" also handles the first line of "ls -l" correctly now.