Crude fix for unterminated escape sequences.
If an escape sequence is longer than BUFSIZ, we'd end up asking read() for zero
bytes.  But if we got zero bytes, we took that as our indication that the child
process had exited and spft would exit as well.  Now we just discard the whole
bufferful if it's just an unterminated escape sequence.  It not a great way to
deal with that (but what is?), but it lets us keep going.

This can happen if you accidentally "cat" a binary file (.jpg's seem
particularly susceptible).